+++ shuup_correios/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# This file is part of Shoop Correios.
+# This file is part of Shuup Correios.
 #
 # Copyright (c) 2016, Rockho Team. All rights reserved.
 # Author: Christian Hess
@@ -7,13 +7,13 @@
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
 
-from shoop.apps import AppConfig
+from shuup.apps import AppConfig
 
 
-class ShoopCorreiosAppConfig(AppConfig):
+class ShuupCorreiosAppConfig(AppConfig):
     name = __name__
-    verbose_name = "Shoop Correios"
-    label = "shoop_correios"
+    verbose_name = "Shuup Correios"
+    label = "shuup_correios"
     provides = {
         "service_provider_admin_form": [
             __name__ + ".forms:CorreiosCarrierForm",
@@ -24,6 +24,6 @@ class ShoopCorreiosAppConfig(AppConfig):
     }
 
 
-default_app_config = __name__ + ".ShoopCorreiosAppConfig"
+default_app_config = __name__ + ".ShuupCorreiosAppConfig"
 
 __version__ = "1.0.0"