+++ shuup_correios/packing/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# This file is part of Shoop Correios.
+# This file is part of Shuup Correios.
 #
 # Copyright (c) 2016, Rockho Team. All rights reserved.
 # Author: Christian Hess
@@ -7,10 +7,10 @@
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
 
-from shoop_correios.packing.base import AbstractPackage, \
+from shuup_correios.packing.base import AbstractPackage, \
     AbstractPackageConstraint, AbstractPackager, BasePackage
 
-from shoop_correios.packing.correios import CorreiosPackage, \
+from shuup_correios.packing.correios import CorreiosPackage, \
     CorreiosPackageConstraint, CorreiosSimplePackager
 
 __all__ = [