+++ shuup_correios_tests/manage.py
@@ -3,7 +3,7 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "shoop_correios.shoop_correios_tests.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "shuup_correios.shuup_correios_tests.settings")
     from django.core.management import execute_from_command_line
     execute_from_command_line(sys.argv)