+++ doc/changelog.rst
@@ -1,4 +1,4 @@
-Shoop Correios Change Log
+Shuup Correios Change Log
 ===================
 
 Version 1.0.0
@@ -6,4 +6,4 @@ Version 1.0.0
 
 Released on 2016-05-09
 
-- The first version of Shoop Correios.
+- The first version of Shuup Correios.